refactor link util code and update unit tests
